split some typeck code into separate modules
